==> code/00_create_folders.R <==
# Run once to create the folder structure 

dir.create('./code')
dir.create('./code/source') #source files location
dir.create('./code/dev') #code that is not finalized and numbered
dir.create('./code/archive') #code that is not used in the final version
dir.create('./data') #post-processed data used in analysis
dir.create('./data/raw') #read-only original data
dir.create('./data/archive') #generated data that are not used in analyses
dir.create('./data/experiments') #generated data that are not used in analyses
dir.create('./results') 
dir.create('./results/figures')
dir.create('./results/figures/poster') 
dir.create('./results/figures/archive') #old figures that are not used
